+++ packages/api-page-builder/src/plugins/models/pbCategory.model.ts
@@ -1,11 +1,19 @@
 import { flow } from "lodash";
 import { validation } from "@webiny/validation";
 
-import { withFields, string, withName } from "@webiny/commodo";
+import { withFields, string, withHooks, withName } from "@webiny/commodo";
 
-export default ({ createBase }) =>
+export default ({ createBase, context }) =>
     flow(
         withName("PbCategory"),
+        withHooks({
+            async beforeDelete() {
+                const { PbPage } = context.models;
+                if (await PbPage.findOne()) {
+                    throw new Error("Cannot delete category because some pages are linked to it.");
+                }
+            }
+        }),
         withFields({
             name: string({ validation: validation.create("required") }),
             slug: string({ validation: validation.create("required") }),