+++ packages/app-headless-cms/src/admin/plugins/fieldRenderers/boolean/index.ts
@@ -1,3 +1,3 @@
 import booleanSwitch from "./booleanSwitch";
 
-export default [booleanSwitch]
+export default [booleanSwitch];