+++ packages/app-plugin-admin-welcome-screen/src/components/Welcome.tsx
@@ -21,6 +21,9 @@ import TwitterIcon from "../icons/twitter-logo.svg";
 import TextbookIcon from "../icons/textbook.svg";
 import LaptopIcon from "../icons/laptop.svg";
 import { AdminWelcomeScreenWidgetPlugin } from "../types";
+import { SecureView } from "@webiny/app-security/components";
+
+import { hasScopes } from "@webiny/app-security"; 
 
 const linkStyle = css({
     textDecoration: "none",
@@ -95,6 +98,21 @@ const Widget = styled("div")({
     marginBottom: "2rem"
 });
 
+const WelcomeScreenWidgetsWrapper = styled("div")({
+    display: "flex",
+    "@media (max-width: 479px)": {
+        flexDirection: "column"
+    }
+});
+
+const cellClass = css({
+    "@media (max-width: 479px)": {
+        "& .webiny-data-list": {
+            margin: "0"
+        }
+    }
+});
+
 const Welcome = () => {
     const security = useSecurity();
 
@@ -104,46 +122,59 @@ const Welcome = () => {
 
     const { fullName } = security.user;
 
+    const canSeeAnyWidget = getPlugins<AdminWelcomeScreenWidgetPlugin>(
+        "admin-welcome-screen-widget"
+    ).some(pl => hasScopes(pl.scopes, { forceBoolean: true }));
+
     return (
         <Grid>
-            <Cell span={12}>
+            <Cell span={12} className={cellClass}>
                 <SimpleForm>
                     <SimpleFormHeader title={`Hi ${fullName.split(" ")[0]}!`} />
                     <SimpleFormContent>
                         <ContentTheme>
                             <Cell span={12}>
                                 <Typography use={"headline6"}>
-                                    <p className={pGetStartedStyle}>
-                                        To get started - pick one of the actions below:
-                                    </p>
+                                    {canSeeAnyWidget && 
+                                        <p className={pGetStartedStyle}>
+                                            To get started - pick one of the actions below:
+                                        </p>
+                                    }
+                                    {!canSeeAnyWidget && 
+                                        <p className={pGetStartedStyle}>
+                                            Please contact administrator for permission to use the site's actions.
+                                        </p>
+                                    }
                                     <br />
                                 </Typography>
                             </Cell>
-                            <div style={{ display: "flex" }}>
+                            <WelcomeScreenWidgetsWrapper>
                                 {getPlugins<AdminWelcomeScreenWidgetPlugin>(
                                     "admin-welcome-screen-widget"
                                 ).map(pl => {
                                     return (
-                                        <Widget key={pl.name}>
-                                            <Elevation z={2} style={{ padding: 10 }}>
-                                                <Typography use={"headline6"}>
-                                                    <p className={widgetTitleStyle}>
-                                                        {pl.widget.title}
-                                                    </p>
-                                                </Typography>
-                                                <Typography use={"body1"}>
-                                                    <p className={widgetDescriptionStyle}>
-                                                        {pl.widget.description}
-                                                    </p>
-                                                </Typography>
-                                                <div className={widgetButtonStyle}>
-                                                    {pl.widget.cta}
-                                                </div>
-                                            </Elevation>
-                                        </Widget>
+                                        <SecureView scopes={pl.scopes}>
+                                            <Widget key={pl.name}>
+                                                <Elevation z={2} style={{ padding: 10 }}>
+                                                    <Typography use={"headline6"}>
+                                                        <p className={widgetTitleStyle}>
+                                                            {pl.widget.title}
+                                                        </p>
+                                                    </Typography>
+                                                    <Typography use={"body1"}>
+                                                        <p className={widgetDescriptionStyle}>
+                                                            {pl.widget.description}
+                                                        </p>
+                                                    </Typography>
+                                                    <div className={widgetButtonStyle}>
+                                                        {pl.widget.cta}
+                                                    </div>
+                                                </Elevation>
+                                            </Widget>
+                                        </SecureView>
                                     );
                                 })}
-                            </div>
+                            </WelcomeScreenWidgetsWrapper>
                         </ContentTheme>
                     </SimpleFormContent>
                     <SimpleFormFooter>
@@ -182,7 +213,11 @@ const Welcome = () => {
                                 <Link to="https://docs.webiny.com/" className={linkStyle}>
                                     <Typography use={"headline5"}>
                                         <div className={footerContainerStyle}>
-                                            <img className={imageStyle} src={TextbookIcon} />
+                                            <img
+                                                className={imageStyle}
+                                                src={TextbookIcon}
+                                                alt={""}
+                                            />
                                             <p className={footerLinkTextStyle}>Documentation</p>
                                         </div>
                                     </Typography>
@@ -199,7 +234,7 @@ const Welcome = () => {
                                 >
                                     <Typography use={"headline5"}>
                                         <span className={footerContainerStyle}>
-                                            <img className={imageStyle} src={LaptopIcon} />
+                                            <img className={imageStyle} src={LaptopIcon} alt={""} />
                                             <p className={footerLinkTextStyle}>Code examples</p>
                                         </span>
                                     </Typography>
@@ -210,25 +245,41 @@ const Welcome = () => {
                                     to="https://github.com/webiny/webiny-js"
                                     className={linkStyle}
                                 >
-                                    <img className={imageStyle} src={GithubIcon} />
+                                    <img
+                                        className={imageStyle}
+                                        src={GithubIcon}
+                                        alt={"Github logo"}
+                                    />
                                     <p>Github</p>
                                 </Link>
                             </Cell>
                             <Cell span={1} className={iconTextStyle} align="middle">
                                 <Link to="https://www.webiny.com/slack/" className={linkStyle}>
-                                    <img className={imageStyle} src={SlackIcon} />
+                                    <img
+                                        className={imageStyle}
+                                        src={SlackIcon}
+                                        alt={"Slack logo"}
+                                    />
                                     <p>Slack</p>
                                 </Link>
                             </Cell>
                             <Cell span={1} className={iconTextStyle} align="middle">
                                 <Link to="https://blog.webiny.com" className={linkStyle}>
-                                    <img className={imageStyle} src={MediumIcon} />
+                                    <img
+                                        className={imageStyle}
+                                        src={MediumIcon}
+                                        alt={"Medium logo"}
+                                    />
                                     <p>Medium</p>
                                 </Link>
                             </Cell>
                             <Cell span={1} className={iconTextStyle} align="middle">
                                 <Link to="https://twitter.com/WebinyPlatform" className={linkStyle}>
-                                    <img className={imageStyle} src={TwitterIcon} />
+                                    <img
+                                        className={imageStyle}
+                                        src={TwitterIcon}
+                                        alt={"Twitter logo"}
+                                    />
                                     <p>Twitter</p>
                                 </Link>
                             </Cell>