+++ packages/ui/src/TopAppBar/index.ts
@@ -5,4 +5,3 @@ export * from "./TopAppBarSection";
 export * from "./TopAppBarTitle";
 export * from "./TopAppBarNavigationIcon";
 export * from "./TopAppBarActionItem";
-